add new blinds type with open/close determination
this fixes the issue that level.blind does not really declare if it is 0=open or 0=closed. Now there are two (hopefully) clear roles that dertermine how the percentage is to be interpreted, i.e. either x% open or x% close(d).
I also added optional tilt using the same method.