fix issues with mediaPlayer
because of array of states it was not detected if no cover was found...

Maybe think about removing the array. It is currently the only pattern that needs it and it requires a lot of stuff later on. Also cover does not show up in devices, so there need to be changes there, too..