add brighness device type
as discussed in #63

I allowed everything ending with `brightness` as role. By looking in the role documentation the only conflicting role *should* be `level.brightness` which *should* be save, because of the `write: false` flag check. Oppinions?

I did so in order to be in line with `humidity` and `temperature` sensor devices which of course is not a hard requirement but nice to have.

Also, if we merge #69, `SharedPatterns.battery` should be added here as well.